Fix running the role against Proxmox hosts
Currently, Ansible doesn't distinguish Proxmox as a separate
Debian-based distribution. Rather, it considers it as a upstream
Debian OS. This requires to divert the package list variables for
Debian 11/12 because of packages being present in Debian but not
in Proxmox and vice-versa.

This commit adds empty package lists for Proxmox 7 (Debian 11) and
Proxmox 8 (Debian 12) to serve as placeholders for future changes.